+++ client/src/components/search.css
@@ -10,8 +10,6 @@ ul {
 };
 
 
-
-
 div.form {
     text-align: left;
     font-size: 30px;
